Move TorchInfoObtainer to separate module
