+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "gpu_utilities"
-version = "0.1.0"
+version = "0.2.0"
 authors = [
 	{ name = "Khambar Dussaliyev", email = "dussaliyev@kolesa.kz"},
 ]
